subway_2/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary # Driver cho PostgreSQL
python-dotenv   # Thư viện để đọc file .env
